23-flashcards_PlacasDeTransito-EmDev/requirements.txt
colorama==0.4.6
customtkinter==5.0.3
darkdetect==0.8.0
Pillow==9.4.0
